+++ src/StochSeq.cpp
@@ -172,7 +172,7 @@ struct StochSeq : Module, Quantize {
 		if (clockTrig.process(inputs[CLOCK_INPUT].getVoltage())) {
 			if (resetMode) {
 				resetMode = false;
-				resetSeqToEnd();
+				resetSeq();
 			}
 			clockStep();
 		}
@@ -245,8 +245,9 @@ struct StochSeq : Module, Quantize {
 		// pitchVoltage = prob * (2 * spread) - spread;
 	}
 
-	void resetSeqToEnd() {
-		gateIndex = seqLength-1;
+	void resetSeq() {
+		// gateIndex = seqLength-1;
+		gateIndex = -1;
 	}
 
 	void invert() {
@@ -605,12 +606,13 @@ struct StochSeqDisplay : Widget {
 		if (layer == 1) {
 
 			// seq position
-			if (module->gateIndex >= 0) {
+			if (module->gateIndex >= -1) {
 				nvgStrokeWidth(args.vg, 2.0);
 				// nvgStrokeColor(args.vg, nvgRGB(128, 0, 219));
 				nvgStrokeColor(args.vg, nvgRGB(0, 238, 255));
 				nvgBeginPath(args.vg);
-				float x = clamp(module->gateIndex * sliderWidth, 0.0, box.size.x - sliderWidth);
+				int pos = module->resetMode ? 0 : clamp(module->gateIndex, 0, NUM_OF_SLIDERS);
+				float x = clamp(pos * sliderWidth, 0.0, box.size.x - sliderWidth);
 				nvgRect(args.vg, x, 1, sliderWidth, box.size.y - 1);
 				nvgStroke(args.vg);
 			}
@@ -628,8 +630,8 @@ struct StochSeqWidget : ModuleWidget {
 
 		StochSeqDisplay *display = new StochSeqDisplay();
 		display->module = module;
-		display->box.pos = Vec(7.4, 86.7);
-		display->box.size = Vec(480, 102.9);
+		display->box.pos = Vec(7.4, 47.7);
+		display->box.size = Vec(480, 141.9);
 		addChild(display);
 
 		addChild(createWidget<JeremyScrew>(Vec(25.9, 2)));
@@ -674,37 +676,37 @@ struct StochSeqWidget : ModuleWidget {
 
 		// addChild(createLight<SmallLight<JeremyBlueLight>>(Vec(84, 50), module, StochSeq::BLUE_LIGHT));
 
-		for (int i = 0; i < NUM_OF_LIGHTS; i++) {
-			float x = 196 * i / NUM_OF_LIGHTS + 5;
+		// for (int i = 0; i < NUM_OF_LIGHTS; i++) {
+		// 	float x = 196 * i / NUM_OF_LIGHTS + 5;
 			
-			float y = ((-std::sin(2.0 * M_PI * i / NUM_OF_LIGHTS) * 0.5 + 0.5) * 50 + 15);
-			// float x = random::uniform() * 260 + 1;
-			// float y = random::uniform() * 50 + 15;
-			// int light = int(random::uniform() * 4);
+		// 	float y = ((-std::sin(2.0 * M_PI * i / NUM_OF_LIGHTS) * 0.5 + 0.5) * 50 + 15);
+		// 	// float x = random::uniform() * 260 + 1;
+		// 	// float y = random::uniform() * 50 + 15;
+		// 	// int light = int(random::uniform() * 4);
 
-			int light = int(i / (NUM_OF_LIGHTS/4.0));
+		// 	int light = int(i / (NUM_OF_LIGHTS/4.0));
 
-			switch(light) {
-				case 0:
-					addChild(createLight<SmallLight<JeremyPurpleLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
-					break;
-				case 1:
-					addChild(createLight<SmallLight<JeremyBlueLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
-					break;
-				case 2:
-					addChild(createLight<SmallLight<JeremyAquaLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
-					break;
-				case 3:
-					addChild(createLight<SmallLight<JeremyRedLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
-					break;
-			}
+		// 	switch(light) {
+		// 		case 0:
+		// 			addChild(createLight<SmallLight<JeremyPurpleLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
+		// 			break;
+		// 		case 1:
+		// 			addChild(createLight<SmallLight<JeremyBlueLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
+		// 			break;
+		// 		case 2:
+		// 			addChild(createLight<SmallLight<JeremyAquaLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
+		// 			break;
+		// 		case 3:
+		// 			addChild(createLight<SmallLight<JeremyRedLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
+		// 			break;
+		// 	}
 
-			// if (random::uniform() < 0.5)
-				// addChild(createLight<SmallLight<JeremyBlueLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
+		// 	// if (random::uniform() < 0.5)
+		// 		// addChild(createLight<SmallLight<JeremyBlueLight>>(Vec(x, y), module, StochSeq::LIGHTS + i));
 
-			// addChild(createLight<SmallLight<JeremyBlueLight>>(Vec(84, 50), module, StochSeq::BLUE_LIGHT));
+		// 	// addChild(createLight<SmallLight<JeremyBlueLight>>(Vec(84, 50), module, StochSeq::BLUE_LIGHT));
 
-		}
+		// }
 
 		addChild(createLight<SmallLight<JeremyPurpleLight> >(Vec(241.1 - 3, 343.2 - 3), module, StochSeq::BANG_LIGHTS + 0));
 		addChild(createLight<SmallLight<JeremyBlueLight> >(Vec(253.7 - 3, 343.2 - 3), module, StochSeq::BANG_LIGHTS + 1));